+++ src/components/sidebarComponents/SidebarComponent.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import ProfileComponent from "./ProfileComponent";
 import HomeIcon from "@mui/icons-material/Home";
@@ -6,6 +7,7 @@ import HistoryIcon from "@mui/icons-material/History";
 import TokenIcon from "@mui/icons-material/Token";
 import SettingsIcon from "@mui/icons-material/Settings";
 import SidebarButtonComponent from "./SidebarButtonComponent";
+import { usePathname } from "next/navigation";
 
 type Props = {};
 
@@ -40,6 +42,12 @@ const buttons = [
 ];
 
 const SidebarComponent = (props: Props) => {
+  const pathname = usePathname();
+
+  if (pathname === "/login" || pathname === "/register") {
+    return null;
+  }
+
   return (
     <div className="flex flex-col max-w-[300px] h-screen border-2 border-[#A5EEA0] bg-white">
       <ProfileComponent name={DEFAULT_NAME} />
