+++ src/models/loginModel.ts
@@ -1,8 +1,4 @@
-
-export interface LoginModel{
-    username : string;
-    password : string;
-
-
+export interface LoginModel {
+  username: string;
+  password: string;
 }
-
